feat: add profile page with loggedin user
