+++ src/components/NavBar/NavBar.css
@@ -2,6 +2,10 @@
   padding: 0 1rem;
 }
 
+.sub-menu {
+  text-decoration: none;
+}
+
 .notification {
   border: 1px solid var(--vanilla);
   background-color: var(--primary-color);
