+++ src/pages/Profile/Profile.css
@@ -38,3 +38,9 @@
 .inverted-text {
   color: var(--primary-color);
 }
+
+/* .profile-page-content .btn .btn-cta {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+} */
