+++ src/pages/Profile/Profile.js
@@ -1,8 +1,21 @@
 import { useAuth } from "../../context";
 import "./Profile.css";
+import { useToast } from "../../custom-hooks/useToast";
+import { useNavigate } from "react-router-dom";
 
 const Profile = () => {
   const { authUser } = useAuth();
+  const { authDispatch } = useAuth();
+  const navigate = useNavigate();
+  const { showToast } = useToast();
+
+  const logoutUser = () => {
+    showToast("Logout Successful", "success");
+    authDispatch({ type: "RESET_AUTH" });
+    localStorage.removeItem("user");
+    localStorage.removeItem("token");
+    navigate("/logout");
+  };
 
   return (
     <div className="profile-page-wrapper">
@@ -28,6 +41,11 @@ const Profile = () => {
               {`${authUser.email}`}
             </div>
           </div>
+        </div>
+        <div>
+          <button className="btn btn-cta" onClick={logoutUser}>
+            Logout
+          </button>
         </div>
       </div>
     </div>
